week2: 영화관 조회 결과값 반환 안되는 오류 해결

+++ spring-boot/src/main/java/com/ceos22/spring_boot/dto/TheaterDto.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 @Getter
 @Builder
 public class TheaterDto {
-    private int theaterId;
+    private Long theaterId;
     private String theaterName;
     private Region region;
     private String address;
